+ Rename file path + Remove console.log + Remove commented code
